[Update]Components folder restrucuring
