+++ src/components/Auth/LoginForm.js
@@ -5,8 +5,8 @@ import { Link } from 'react-router-dom';
 import RaisedButton from 'material-ui/RaisedButton';
 import { TextField } from 'redux-form-material-ui';
 
-import { login } from '../actions/authActions';
-import { email, required } from '../helpers/formValidators';
+import { login } from '../../actions/authActions';
+import { email, required } from '../../helpers/formValidators';
 
 const style = {
   margin: 12,
