+++ src/components/Items/ItemEdit.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 
-import ShoppinglistForm from './ShoppinglistForm';
-import { editShoppinglist, fetchShoppinglist } from '../actions/shoppinglistActions';
+import ItemForm from './ItemForm';
+import { editItem, fetchItem } from '../../actions/itemActions';
 
-class ShoppingListEdit extends Component {
+class ItemEdit extends Component {
   constructor(props) {
     super(props);
 
@@ -12,20 +12,20 @@ class ShoppingListEdit extends Component {
   }
 
   componentDidMount() {
-    const { id } = this.props.match.params;
-    this.props.fetchShoppinglist(id);
+    const { shId, itId } = this.props.match.params;
+    this.props.fetchItem(shId, itId);
   }
 
-  onSubmit(name) {
-    const { id } = this.props.match.params;
-    this.props.editShoppinglist(id, name, (path) => {
-      this.props.history.push(path);
-    });
+  onSubmit(name, quantity) {
+    const { shId, itId } = this.props.match.params;
+    this.props.editItem(shId, itId, name, quantity);
+    this.props.history.push(`/shoppinglist/${shId}`);
   }
 
   render() {
-    const { editingShoppinglist, shoppinglist } = this.props;
-    if (!shoppinglist) {
+    const { editingItem, item } = this.props;
+    const { shId } = this.props.match.params;
+    if (!item) {
       return (
         <div>
           Loading...{' '}
@@ -36,24 +36,25 @@ class ShoppingListEdit extends Component {
         </div>
       );
     }
-    const { name } = shoppinglist;
-
+    const { name } = item;
+    const quantity = parseFloat(item.quantity);
     return (
       <div className="form-container">
-        <ShoppinglistForm
-          initialValues={{ name }}
+        <ItemForm
+          initialValues={{ name, quantity }}
           onSubmit={this.onSubmit}
-          editingShoppinglist={editingShoppinglist}
+          editingItem={editingItem}
           context="Edit"
+          shId={shId}
         />
       </div>
     );
   }
 }
 
-const mapStateToProps = ({ shoppinglists, editingShoppinglist }, ownProps) => ({
-  editingShoppinglist,
-  shoppinglist: shoppinglists[ownProps.match.params.id],
+const mapStateToProps = ({ items, editingItem }, ownProps) => ({
+  editingItem,
+  item: items[ownProps.match.params.itId],
 });
 
-export default connect(mapStateToProps, { editShoppinglist, fetchShoppinglist })(ShoppingListEdit);
+export default connect(mapStateToProps, { editItem, fetchItem })(ItemEdit);
