+++ src/components/UI/Welcome.js
@@ -4,7 +4,10 @@ import { Card, CardMedia } from 'material-ui/Card';
 const Welcome = () => (
   <Card>
     <CardMedia>
-      <img src="https://www.buzzfeed.com/static-assets/img/shopping_header.bd16ddeece64859c9f790c30a332a0ed.png" alt="" />
+      <img
+        src="https://www.buzzfeed.com/static-assets/img/shopping_header.bd16ddeece64859c9f790c30a332a0ed.png"
+        alt=""
+      />
     </CardMedia>
   </Card>
 );
